chore(deps): Update Deta library version to 0.0.3
The Deta library version in the project dependencies has been updated
to 0.0.3. This update ensures that the latest features and bug fixes
provided by the library are available for use in the project.

For alternative installation methods, such as Clojure CLI/deps.edn,
Gradle, and Maven, please refer to the Installation section of the
documentation for detailed instructions.