chore(docs): Update README section order
This commit reorganizes the sections in the README.md file to improve
clarity and consistency. The 'Updating Data' and 'Coverage' sections were
incorporated to better structure the document, providing a more logical
order for users to find information about updating data and test coverage
in the project.
- The 'Updating Data' section was added to showcase examples of updating
data in the Deta Base.
- The 'Coverage' section was included to explain how to generate coverage
reports using the 'lein cloverage' command.
- The order and content changes aim to enhance user experience and
navigation through the document.